Render email templates using FreeMarker (#258)
Replace the not-really-a-template-system string replacement in `EmailService` with
Apache FreeMarker. This will give us more templating flexibility as we introduce
additional kinds of email notifications.

The templates for the plaintext email bodies now have `.txt.ftl` extensions so
that editors will detect them as plaintext FreeMarker templates and give us syntax
highlighting and so forth. The templates for HTML email bodies, which are still
generated from MJML, now have `.ftlh` extensions which causes FreeMarker to treat
them as HTML and automatically enable escaping of HTML special characters.