SW-893 Remove tracking features (#279)
We aren't going to ship our original concept of the plant tracking features.
Remove its implementation from the code base.

Nearly all of this is just straightforward deletion of code. One minor
exception is that this _adds_ a previously-missing field to `SitesTable` so we
can continue to test that the search code returns geometry values in the
correct format.