Use model classes to pass values to email templates
We're going to be adding a number of email templates soon. To reduce the chances
that we'll inadvertently mistype the name of a template variable, switch from
generic `Map` objects to model classes to hold the values the template can access,
and annotate the template files so IntelliJ can do autocompletion and error
checking.