SW-856 Move email header/footer to include files
To avoid having to include the boilerplate header and footer content in every
email template, move them to separate MJML files that get included by the
MJML files for individual notification types.

Since the header and footer need to know the base URL of the web app (because
they use it to construct URLs of the company logo and other images), move it
to `EmailTemplateModel` and turn that from a marker interface to a class.