SW-916 Support search filters on compound primary keys (#293)
Previously, the search table definitions for the `organization_users` and
`project_users` tables treated user ID as the primary key. But that caused a
problem: since search criteria are implemented using an `IN` clause that
selects the top-level objects by primary key, the results weren't being filtered
by organization, just by user.

So for example, if you searched for organization members with a role of "Admin"
in a particular organization ID, you would only get search results for users who
had the admin role in that organization, but you would get the membership details
for _all_ their organizations, not just the one you specified.

The fix is to not ignore the fact that organization users (and project users) have
compound primary keys.

Luckily, jOOQ has a feature called "embeddable types" that lets us solve this
without having to revamp our search implementation: you can define synthetic data
types that combine multiple columns into a single field that acts, from the
application's point of view, like a single atomic value. Configuring the search
tables to use these embeddable types as the primary key, rather than the user ID,
causes the filtering to be applied to both the organization and user at the same
time, and the queries return the expected results.