Prevent overlapping GBIF imports
Previously, if you double-clicked the submit button in the admin UI's
GBIF import form, the server would run both imports at the same time.
This could result in duplicate data being imported.

Add a locking mechanism so the second attempt will fast-fail before
making any changes.

The locking mechanism is implemented as a general-purpose service since
this is unlikely to be the last time we'll need application-wide locking.