Allow uploading GBIF backbone zipfile to admin UI (#311)
The existing API endpoint for importing GBIF backbone data requires unpacking
the zipfile from the GBIF website and then constructing a correct URL to point
to the location of the unpacked files.

That's workable, but kind of a hassle, especially in dev environments.

Add a section to the admin UI to allow uploading the original zipfile without
having to unpack it or construct a URL for it. (I chose to do this as a file
upload rather than having the server directly download from the GBIF site
because we'll need to import the data over and over again in dev environments
and it's not nice to eat GBIF's bandwidth repeatedly downloading the same big
file.)

Importing species data is only permitted for super-admin users. There wasn't
previously any documentation describing how to set someone as a super-admin,
so add a section to `README.md` about that.