Support unknown-length files in FileStore
Previously, mostly due to the fact that we were using the simple `PutObject` S3
API, callers had to tell `FileStore.write()` the length of the file they wanted
to store.

Unfortunately, that doesn't play well with streaming uploads, where the incoming
request doesn't specify a content length.

Change `S3FileStore` to use the S3 multipart upload API, which supports uploading
a series of chunks and then explicitly marking the upload as complete after the
last chunk is written.

With that change in place, remove the `size` argument from `FileStore.write()`
and update callers accordingly.