SW-909 Data model and permissions for uploaded files (#333)
Add tables to represent files that are uploaded by users and processed
asynchronously by the server.

This doesn't include any logic specific to species list CSV uploads; it's just the
file-type-agnostic scaffolding that we can also use for other kinds of files
(accession lists, etc.) down the road.

Currently, the permissions model is simply, "You have permission to do anything
to your own uploads, but no permission to do anything with anyone else's." That's
what we want for the current product requirements; down the road we might want to
allow organization admins to work with their users' uploaded files, so this change
adds the usual fine-grained permission methods to let us do that later on.