Add client for Balena API
In preparation for syncing device manager status information from Balena, add a
client that can make the necessary Balena API requests.

It must be explicitly enabled in the server configuration, and the configuration
must include a valid API key and the ID of an existing Balena fleet. If it's not
enabled, terraware-server will not attempt to communicate with Balena.