SW-999 Stop autocreating seed banks for new orgs
Now that we're adding the ability for users to manage their own seed banks, there
is no longer a need to automatically create a seed bank when a user creates a new
organization.

However, since the autocreated seed bank project is special (it is configured as
an organization-wide project) and we don't currently allow clients to create that
kind of project, we still want to autocreate the project and the site that clients
will use when they create new seed bank facilities.