SW-1009 Add device managers and devices to admin UI
Update the admin UI to include device managers and devices. The initial use case
here is dev environments: developers can create device managers locally to test
the sensor kit setup workflow. It thus doesn't include the ability to view and
edit absolutely everything about devices, just the fields that will be needed
for sensor kit setup testing.

The basic flow will look something like:

1. Create a new seed bank facility, either via API or in the admin UI.
2. Create a new device manager in the admin UI's device managers page.
3. Connect the device manager to the facility, either via API or by editing the
   device manager's facility ID in the admin UI. (The behavior is a a little
   different: the admin UI won't create a Keycloak user.)
4. Add devices from the admin UI's facility page. There are convenience links
   to prefill the device settings for a couple different device types.

The device manager page can be used to manipulate some values that will come
from Balena in the real implementation. The two most useful are the "Online?"
setting and the "Update Progress" percentage.

There's currently no way to disconnect a device manager from a facility since it's
not a product requirement and would involve figuring out what the behavior should
be (e.g., what happens to the Keycloak user?) In dev environments, it should be
simple enough to create a new facility and device manager to test with.