SW-1007 Add device managers to data model and API
Add a table for tracking device manager information. The table has a
combination of locally-maintained data and data from the Balena API, though the
actual Balena API calls are not included in this change.

In addition, the facilities table gains a "connection state" column that indicates
whether or not a device manager has been associated with the facility. This is
included in the facility information returned by `/api/v1/organizations` and
`/api/v1/facilities`. A later change will add a way for clients to tell the server
that device maanger configuration is finished.

`GET /api/v1/devices/managers` allows searching for device managers by short code.
The search is currently case-sensitive; we don't know what the short codes are
going to actually look like yet, so it's unclear if they should be case-sensitive.

`POST /api/v1/devices/managers/{id}/connect` allows connecting a device manager to
a facility. This also creates a new Keycloak user and updates the connection state
of the facility. A later change will also make this configure the device via the
Balena API, but for now there's no Balena interaction.

One thing of note here is the way permissions work. Device managers are an unusual
case in that they need to be visible to users before the system knows which
organizations own them. So the permission logic here is basically that a device
manager is visible to all users if it's not connected yet, but then once it's
connected, it's only visible to users who have permission to see its assigned
facility.