SW-1021 Add endpoint to list devices' timeseries
Previously, there was no way for a client to discover which timeseries existed for
a device, just to probe whether a timeseries with a specific name existed.

Add an endpoint to list the timeseries for one or more devices.

The response payload includes the latest value for each timeseries, if any,
including its timestamp.

The immediate use case for this is a tool to generate dummy timeseries data, but
it should be generally useful for any client that needs to show a list of
available timeseries.