SW-1021 Script to create random timeseries data (#369)
Add a script to populate a terraware-server database with random timeseries data
that is in the appropriate ranges for specific kinds of devices.

In dev environments, you can follow a workflow like:

1. Create a seed bank facility (via API or the admin UI)
2. Connect a device manager to it using the admin UI
3. Create PV and temperature sensor devices using the admin UI
4. Mark the facility as configured (via API or the admin UI)
5. Run the new script to fill up the database with a month's worth of data

The script will probably need to be improved as we start using it in earnest for
development, but should be full-featured enough to be useful already.