SW-1034 Automatically create automations
The device manager needs to know the acceptable ranges of various timeseries so it
can generate notifications if they go out of range.

The correct ranges, and the relevant timeseries, depend on various attributes of
the device, including its name. In the future, we'll likely also want to let users
edit the ranges on a per-device basis. So the server needs to be the source of
truth for the ranges.

Add logic to create and remove automations based on changes to device data.