Move email template directory names to model classes
To prepare for device alerts where we need to use different email messages
depending on the specific kind of device and the specific timeseries, move the
email template subdirectory names to the template model classes. This isn't
textbook clean MVC, but will keep the logic for choosing templates based on
the model data close to the model data rather than spreading it around the
code base.