SW-955 API support for species problems
Update the existing species endpoints to include information about the results of
name checking.

The new `POST /api/v1/species/problems/{id}` endpoint applies the suggested change
to the species name, if any. `DELETE /api/v1/species/problems/{id}` removes the
problem report. If the user subsequently edits the species name, the system will
check it again, but if the user edits other information about the species, any
previously-deleted problems won't be flagged again. (That is, we honor the user's
request to ignore a particular problem.)