SW-1051 Make test clock work in clustered environments (#411)
Previously, the time offset for the test clock was loaded once at server start
time and then only updated if an admin wanted to adjust the clock. This is fine in
dev environments, but doesn't work well in a clustered deployment such as staging,
where there are multiple server hosts; they would disagree about what time it was
supposed to be.

Add a timer task to poll the offset every 5 seconds.

Any business logic that gets executed when the clock is adjusted will only happen
on the host that's processing the admin request to update the time.