SW-1064 Disable unwanted health checks (#413)
Spring Boot has a built-in component called Actuator. It sets up health checks
based on what components the application is using, e.g., if you're using a
database, it will make sure the database is reachable. We want most of the
default checks, but two of them are not useful for us: the free disk space check
(which wouldn't work right from inside a Docker container) and the mail server
connectivity check (we don't want to mark terraware-server as down if AWS's mail
service is having an outage).