SW-1053 Apply previous name changes on species CSV import (#415)
If the user uploads a species list CSV then changes the scientific name of one of
the species from the CSV (e.g., by accepting a suggested name change from the
app), we want to remember the name change if they upload their species list again.

Add an "initial scientific name" to the species table where we track the name that
was used when the species was first created. Species CSV validation and import now
looks at the initial name in addition to the current name, and will attempt to
update the existing species if the CSV still lists it with its original name.