SW-1059 Rename short code to sensor kit ID
We decided to use the term "sensor kit ID" as the official name for the ID code
that gets printed on sensor kits and that users enter to connect a sensor kit to a
seed bank, so update the code to use the official terminology.

This affects the data model, the implementation, and the API. For now, the API
continues to also accept `shortCode` parameters in requests and continues to
include `shortCode` values in responses, alongside `sensorKitId`, to give clients
a migration path. We'll follow up and remove support for `shortCode` once the
client code has shifted to the new name.