SW-1111 Make text field searches case-insensitive (#420)
Previously, the search API did case-sensitive matching for `Exact` search
filters on text fields and was inconsistent about case sensitivity for `Fuzzy`
searches based on the length of the search string. The latter was a side effect
of the fact that we were using a particular fuzzy match operator from the
`pg_trgm` extension.

Change it to always do case-insensitive searches and to use a more relaxed
fuzzy match operator that supports shorter search strings.