SW-1117 Filter deleted species out of search results
Previously, the search API wasn't paying attention to whether or not a species was
deleted. One option is to let the client decide whether or not to include deleted
species. But users shouldn't be able to see deleted species at all, so instead,
treat it as a permissions filter that always gets applied.