SW-1016 Fix jOOQ forced type warning
We generate an enum for the `notification_criticalities` table, but that table is
only ever referenced by the `notification_types` table, which itself gets turned
into an enum.

For each of our generated enums, we use a "forced type" to tell jOOQ to use it
instead of the underlying integer type when it generates the Kotlin
representation of a table. But in the case of `notification_criticalities`,
jOOQ never generates any code that references it, which causes jOOQ to emit a
warning about the type being unused.

There doesn't seem to be any straightforward way to automatically detect that
situation in such a way that we wouldn't also suppress legitimate warnings when
we misspell a table or column name in the configuration. So add an explicit
config option to turn the forced type generation off.