SW-1026 Update settings in devices table
Currently, the device manager always uses a device setting called `diagnosticMode`
to control the logging verbosity level for a device. Pull that value up into the
standard device schema.

For now, the verbosity level is returned both as a top-level payload field and as
a settings value called `diagnosticMode`, though it is no longer stored in the
settings JSON blob in the database. Once the device maanger is updated to use the
`verbosity` field, this backward-compatibility code can be removed.

The device manager doesn't use the "polling interval" value at all (the interval
depends on the device type and isn't user-adjustable), so remove that from the
table and the API.