Remove support for non-PostgreSQL fuzzy text search (#432)
Previously, fuzzy text search was implemented in such a way that it would be
possible for the code to support more than one database engine. But we've
decided to only target PostgreSQL, so there's no need for the added indirection.

Turn the `likeFuzzy` and `similarity` methods into top-level extension methods
and get rid of the classes that used to be needed to support runtime-selectable
implementations.