SW-1049 Make AccessionStore.fetchOneById non-nullable
Rename `fetchById` to `fetchOneById` for consistency with other store classes.

All the call sites were throwing exceptions on failure anyway, so consolidate the
exception throwing into the fetch method and make the call sites less cluttered.
This also eliminates the need for the "skip permission check" parameter which was
always kind of screwy.