SW-1128 Generate report of dependency licenses (#436)
In preparation for releasing the server as open source, add documentation of
the licenses of all its dependencies. This will make it easier to confirm that
the license we choose is compatible with the licenses of all our dependencies.

The license report can be generated using `./gradlew generateLicenseReport`
and will be automatically generated and published to the repo's GitHub Pages
when dependencies are modified.