SW-951 Add authentication to OpenAPI schema (#437)
Update the OpenAPI schema to include information about how users can send
authenticated API requests.

Configure the Swagger UI to pass session cookies through to the backend. This
will allow us to use the Swagger UI's "try it out" feature to send API requests to
the server, so long as we've previously logged into the web app and have a valid
session cookie.