SW-965 Deleting notifications after a configured number of days
Allow product to be able to clean up notifications, current default for expiration of notifications is 30 days.
Created a NotificationCleanupTask that piggy backs onto DailyTaskRunner's DailyTaskArrivedEvent.
Wondering if maybe this should be its own cron job that runs at a higher frequency.
Tested on my local dev server setup and adjusting the clock in the admin interface.