SW-1049 Make ProjectStore.fetchOneById non-nullable (#445)
Move the "throw exception if the organization isn't readable" logic from the call
sites to the fetch method, and rename it from `fetchById` to `fetchOneById` for
consistency with other store classes. Also rename `fetchAll` to `findAll` for the
same reason.

This requires adding project-read permission to `AppNotificationServiceTest`, but
the new permission is consistent with the existing ones in that test.