Improve performance of /api/v1/species/lookup/details
This endpoint does an exact name lookup on a species to get additional information
about it from the GBIF database, e.g., its family name and common names.

On PostgreSQL 14 and higher, it uses the trigram index on `gbif_names.name`. But
using trigram indexes for exact lookups isn't supported on PostgreSQL 13, which we
are running in prod/staging. So the endpoint has to fall back to a full scan of
`gbif_names`. That scan isn't disastrously slow, but it's noticeably slower than
an index lookup.

Add a B-tree index to support exact lookups on earlier PostgreSQL versions.

Even after we upgrade the prod/staging databases, we'll probably still want to
keep this index since trigram index lookups are slower than B-tree lookups for
exact values.