SW-1061 Add unit tests for notification email
Add a test suite to confirm that the rendered email notification messages
include the expected content. This doesn't test the boilerplate text in the
template files since those are subject to change, but it attempts to verify
that the links are present and pointing to the right place and that the various
bits of dynamic content specified in the requirements docs are included.

It is still possible for someone to mess up an MJML template and cause the
notifications to look illegible, so when we revise notification content, we
will still want to do manual testing to confirm the results have the correct
appearance. But this set of tests will at least confirm that the templates
have no syntax errors and use the expected set of variable substitutions.

Writing this test suite turned up a bug: the notification for unknown
automations wasn't linking to the specific device on the monitoring dashboard.
That bug is fixed here.