Analyze GBIF tables after import
It seems PostgreSQL doesn't reliably autoanalyze tables after big bulk
delete/insert operations, so it was trying to plan queries on the GBIF tables
without the benefit of statistics about the distribution of values in various
columns.

Currently, someone has to manually run `ANALYZE` on all the tables after an
import, but that's easy to forget to do.

Make the code automatically do it at the end of the import process.