Move custom tasks out of build.gradle.kts
Our Gradle build script is a little hard to navigate in part because it has a
bunch of one-off custom tasks to implement custom build logic. Turn these into
full-fledged classes that use Gradle's task API. This should declutter the build
script a bit and also make it easier to improve the custom logic in the future.