SW-1183 Add facilities to organization payload (#464)
In preparation for removing projects and sites from our data model, add a
`facilities` field to the response payloads of `/api/v1/organizations` and
`/api/v1/organizations/{id}` that contains a list of all the facilities the
user has access to. It is populated if the `depth` query parameter is set to
`Facility`.

The facilities continue to also be included under sites so as not to break
existing clients.