SW-292 Add DeviceManagerUser with restricted permissions
Device managers don't need access to the same set of operations as any kind of
human user, and their association with organizations and facilities works
differently than it does for human users, so creating them as regular users with
a particular role isn't a great approach.

Introduce a new `TerrawareUser` implementation that's specifically for device
manager users and grants them exactly the permissions they need.