SW-292 Remove API key management from admin UI
Originally, we had the idea that we'd be offering up a public API to customers and
letting them write their own client code. So the code has a concept of an "API
client" as a special kind of user, and the admin UI lets you create those users
and generate authentication tokens for them.

But we haven't ended up going that direction; instead we've ended up piggybacking
on the "API client" feature set for device manager authentication but not using it
for anything else.

Remove API key management from the admin UI.