Remove redundant features from admin UI
The admin UI doesn't need to support doing things that you can already do in the
customer-facing web app.

Remove all the user management features (these are available via the People page
in the web app) and the ability to create an organization (which can be done via
the web app's navbar).

Creating and editing facilities is still supported since the admin UI exposes
some settings that aren't available in the web app.