SW-292 Reduce permissions for non-admin users (#486)
Previously, device manager users were assigned the Contributor role which meant
that that role had to be able to perform all the operations a device manager uses.

Now that there's a separate user type for device managers, we no longer need to
grant access to operations that a regular non-admin user has no business doing.

This only applies to Contributor and Manager user types; Owner and Admin roles
continue to grant access to device-related operations.