SW-1124 Add viability test as alternate name in API
Update the API payload definitions to allow clients to use either "germination
test" or "viability test" in various field names, and to include both names in
responses. If both are present, "viability test" takes precedence.

The search API is _partially_ updated here, to add alternate names for some
existing sublists and a field. The old names continue to work.

* `accessions.germinationTests` -> `accessions.viabilityTests`
* `accessions.latestGerminationTestDate` -> `accessions.latestViabilityTestDate`
* `germinationTests.germinations` -> `viabilityTests.viabilityTestResults`
* `germinations.germinationTest` -> `viabilityTestResults.viabilityTest`

There are some field aliases in the `accessions` search table that still use the
old names, and this change doesn't add equivalents for them. Instead, we'll want
to update the client to use the canonical field names, after which we can get
rid of the aliases entirely.