SW-1124 Rename "germination test" to "viability test" (#489)
* SW-1124 Rename "germination test" to "viability test"

In preparation for updating the seed management code, bring the terminology we use
internally closer to the user-facing terminology by changing the term "germination
test" to "viability test."

This is only an update to the names the server uses internally. The API continues
to use the existing field names, and the term "germination test" still appears as
a value in some enums because they are included in API payloads and we don't want
to use different values the client isn't ready for. A subsequent change will update
the API to accept both kinds of names, and then a third change will remove support
for the old names.

Rename the following tables and their corresponding classes in the code:

* `accession_germination_test_typess` to `accession_viability_test_types`
* `germinations` to `viability_test_results`
* `germination_seed_types` to `viability_test_seed_types`
* `germination_substrates` to `viability_test_substrates`
* `germination_tests` to `viability_tests`
* `germination_test_types` to `viability_test_types`
* `germination_treatments` to `viability_test_treatments`

Rename fields in model classes to reflect the new terminology.

Rename viability-test-related methods and local variables in various classes.

Rename the `germinationTest` MJML template directory to `viabilityTest`.

Update references to "germination test" in comments, log messages, and test names
to use "viability test" instead.

All the code renames were done using IntelliJ "rename" refactoring. In cases
where the code had fields or parameters in alphabetical order, the order was
updated using IntelliJ "change method signature" refactoring.

* Rename some constraints and indexes I missed earlier; rename a notification type since notification types are not exposed as part of the API