SW-1260 Remove accession status notifications
Currently, the app sends per-accession notifications for various automated
status transitions as well as notifications that summarize the statues of all the
accessions at a seed bank.

We are simplifying the workflow, removing some of the transitions, and we are also
making many of the transitions manual rather than automated, so these
notifications will shortly become irrelevant.

Keep the per-accession notification for "scheduled drying end date has arrived"
since we are still going to want that when we move to the new set of statuses.